LayoutTests/inspector/debugger/dynamic-script-tag.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/debugger-test.js"></script>
<script src="../../http/tests/inspector/workspace-test.js"></script>
<script>
document.write("<scrip" + "t>function foo() { }</sc" + "ript>");
var scriptElement = document.createElement("script");
scriptElement.textContent = "function bar() { }";
document.head.appendChild(scriptElement);

var test = function()
{
    InspectorTest.startDebuggerTest(step1);

    function step1()
    {
        var model = WebInspector.debuggerModel;

        var scripts = Object.values(WebInspector.debuggerModel.scripts);
        function filterOutStandaloneScripts(script)
        {
            return script.sourceURL.indexOf("dynamic-script-tag.html") !== -1;
        }
        scripts = scripts.filter(filterOutStandaloneScripts);

        function compareScriptFunction(script1, script2)
        {
            return script1.sourceURL.localeCompare(script2.sourceURL);
        }
        scripts.sort(compareScriptFunction);
        InspectorTest.assertEquals(3, scripts.length, "Wrong scripts count.");

        var uiSourceCodes = [];
        InspectorTest.addResult("Dumping scripts:");
        for (var i = 0; i < scripts.length; ++i) {
            var script = scripts[i];
            InspectorTest.addResult("script" + (i+1) + ": " + script.sourceURL);
            InspectorTest.addResult("Is anonymous:" + script.isAnonymousScript());
            InspectorTest.addResult("Is inline script:" + script.isInlineScript());
            if (!script.isInlineScript())
                uiSourceCodes.push(script.rawLocationToUILocation(0, 0).uiSourceCode);
        }

        function compareUISourceCodeFunction(uiSourceCode1, uiSourceCode2)
        {
            return uiSourceCode1.url.localeCompare(uiSourceCode2.url);
        }
        uiSourceCodes.sort(compareUISourceCodeFunction);

        var seenUISourceCodes = new Map();
        InspectorTest.assertEquals(2, uiSourceCodes.length, "Wrong uiSourceCodes count.");
        InspectorTest.addResult("Dumping uiSourceCodes:");
        
        var index = 0;
        dumpNextUISourceCode();
        function dumpNextUISourceCode()
        {
            if (index >= uiSourceCodes.length) {
                uiSourceCodesDumped();
                return;
            }
            var uiSourceCode = uiSourceCodes[index];
            InspectorTest.dumpUISourceCode(uiSourceCode, uiSourceCodeDumped);

            function uiSourceCodeDumped()
            {
                InspectorTest.addResult("UI resource: " + uiSourceCode.resource());
                InspectorTest.assertTrue(!seenUISourceCodes.get(uiSourceCode), "UISourceCode matches two scripts: " + uiSourceCode.url);
                seenUISourceCodes.put(uiSourceCode, uiSourceCode);
                ++index;
                dumpNextUISourceCode();
            }
        }
        
        function uiSourceCodesDumped()
        {
            InspectorTest.completeDebuggerTest();
        }
    }
}
</script>
</head>
<body onload="runTest()">
<p>Tests that inline scripts and document.write scripts get different uiSourceCodes with different URLs.
<a href="https://bugs.webkit.org/show_bug.cgi?id=87119">Bug 87119</a>
</body>
</html>
